medimind-backend/requirements.txt
fastapi
uvicorn[standard]
motor
redis
passlib
bcrypt
python-multipart
pydantic
pydantic-settings
pymongo
easyocr
openai
python-dotenv
apscheduler
